+++ src/session_timers.rs
@@ -1,6 +1,9 @@
 use chrono::{DateTime, Utc};
 
-///Internal Timers to keep track of when last ran an expiry or database sweep.
+/// Internal Timers
+///
+/// used to keep track of the last ran expiration check for both database and memory session data.
+///
 #[derive(Debug)]
 pub struct AxumSessionTimers {
     pub last_expiry_sweep: DateTime<Utc>,