+++ CHANGELOG.md
@@ -6,7 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 # Unreleased
 
-- None.
+- **breaking:** GDPR Compliance.
+- **breaking:** Data Cookie ID only set if disable_gdpr is true or Accepted is true.
+- **added:** Config for GDPR Sessions.
+- **added:** GDPR Memory and Database Session unloading on not accepted.
+- **added:** Accepted GDPR Cookie.
+- **breaking:** GDPR is Enabled by default so you must use set_accepted on user data for a session to save or disable gdpr_mode.
+- **added:** Better overall documentation with Doc Tests.
 
 # 1.1.0 (12. May, 2022)